feat: add condition to throw error if someone try to remove a user with dungeon_master role

+++ src/use-cases/parties/members/remove/remove-member.ts
@@ -15,6 +15,10 @@ export class RemoveMemberUseCase {
       throw new ResourceNotFoundError()
     }
 
+    if (member.data.role === 'dungeon_master') {
+      throw new Error('Dungeon master cannot be removed from the party')
+    }
+
     await this.partyMembersRepository.delete({
       partyId: data.partyId,
       userId: data.userId,
